fix some size_t errors for length. Need int for std::string constructor.
git-svn-id: svn://136.177.114.72/svn_GW/phastpp/branches/phastpp3@8333 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phast/PhreeqcRM/PhreeqcRM.cpp
@@ -1599,7 +1599,7 @@ PhreeqcRM::Char2TrimString(const char * str, size_t l)
 	{
 		if (l > 0)
 		{
-			std::string tstr(str, l);
+			std::string tstr(str, (int) l);
 			stdstr = tstr;
 		}
 		else
